requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
comtypes==1.4.16
cryptography==46.0.7
distro==1.9.0
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
markdown-it-py==4.0.0
mdurl==0.1.2
MouseInfo==0.1.3
openai==2.32.0
pillow==12.2.0
playwright>=1.49.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==2.0.1
pyperclip==1.11.0
PyAudio==0.2.14
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.99
python-dotenv==1.2.2
pytweening==1.2.0
pywin32==311
pywinauto==0.6.9
PyYAML==6.0.3
SpeechRecognition==3.16.0
requests==2.33.1
rich==15.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
